requirements.txt
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.4
pytest>=7.0.0
pytest-cov>=3.0.0
